refactor: move DNS resolver
